extract color hash function to a mixin
