+++ resources/assets/js/admin/components/slides/annotations/AnnotationsList.vue
@@ -104,6 +104,8 @@
 </style>
 
 <script>
+	import string_color from 'js/admin/mixins/string-color'
+
 	export default {
 		name: 'AnnotationsList',
 		data() {
@@ -121,27 +123,8 @@
 				default: 0
 			}
 		},
+		mixins: [ string_color ],
 		methods: {
-			getColourForStr(str) {
-				let hash = 0;
-				for (let i = 0; i < str.length; i++) {
-					hash = str.charCodeAt(i) + ((hash << 5) - hash);
-				}
-
-				let hex =
-				((hash >> 24) & 0xff).toString(16) +
-					((hash >> 16) & 0xff).toString(16) +
-					((hash >> 8) & 0xff).toString(16) +
-					(hash & 0xff).toString(16);
-
-				hex += "000000";
-				hex = hex.substring(0, 6)
-				const r = parseInt(hex.substring(0,2), 16);
-				const g = parseInt(hex.substring(2,4), 16);
-				const b = parseInt(hex.substring(4,6), 16);
-
-				return `rgba(${r}, ${g}, ${b}, 0.2)`
-			},
 			isEven(index) {
 				return index % 2 === 0
 			},
